zng_wgt_fill
Module node

Color and gradient fill nodes and builders.

Structs

ConicGradient

Conic gradient.

GradientBuilder

Gradient builder start.

LinearGradient

Linear gradient.

RadialGradient

Radial gradient.

TiledConicGradient

Tiled conic gradient.

TiledLinearGradient

Tiled linear gradient.

TiledRadialGradient

Tiled radial gradient.

Functions

conic_gradient

Starts building a conic gradient with the angle and color stops.

flood

Node that fills the widget area with a color.

gradient

Starts building a gradient with the color stops.

linear_gradient

Starts building a linear gradient with the axis and color stops.

radial_gradient

Starts building a radial gradient with the radius and color stops.