Function zng_wgt_fill::node::conic_gradient

pub fn conic_gradient<C, A, S>(
    center: C,
    angle: A,
    stops: S
) -> ConicGradient<S::Var, C::Var, A::Var, LocalVar<ExtendMode>>
where
    C: IntoVar<Point>,
    A: IntoVar<AngleRadian>,
    S: IntoVar<GradientStops>,

Starts building a conic gradient with the angle and color stops.

Returns a node that is also a builder that can be used to refine the gradient definition.