zng_wgt_fill::node

Function linear_gradient

pub fn linear_gradient(
    axis: impl IntoVar<LinearGradientAxis>,
    stops: impl IntoVar<GradientStops>,
) -> LinearGradient

Starts building a linear gradient with the axis and color stops.

Returns a node that is also a builder that can be used to refine the gradient definition.