zng_wgt_fill::node

Struct RadialGradient

pub struct RadialGradient { /* private fields */ }

Radial gradient.

Can be used as a node that fills the available space with the gradient, or can continue building a radial or tiled radial gradient.

Use gradient or radial_gradient to start building.

Implementations

impl RadialGradient

pub fn extend_mode(self, mode: impl IntoVar<ExtendMode>) -> RadialGradient

Sets the extend mode of the radial gradient.

By default is ExtendMode::Clamp.

pub fn repeat(self) -> RadialGradient

Sets the extend mode to ExtendMode::Repeat.

pub fn reflect(self) -> RadialGradient

Sets the extend mode to ExtendMode::Reflect.

pub fn tile( self, tile_size: impl IntoVar<Size>, tile_spacing: impl IntoVar<Size>, ) -> TiledRadialGradient

Continue building a tiled radial gradient.

pub fn tile_size(self, size: impl IntoVar<Size>) -> TiledRadialGradient

Continue building a tiled radial gradient.