zng_wgt_fill::node

Struct TiledLinearGradient

pub struct TiledLinearGradient<S, A, E, O, T, TS> { /* private fields */ }

Tiled linear gradient.

Can be used as a node that fills the available space with the gradient tiles, or can continue building a tiled linear gradient.

Use LinearGradient::tile to build.

Implementations

impl<S, A, E, O, T, TS> TiledLinearGradient<S, A, E, O, T, TS>
where S: Var<GradientStops>, A: Var<LinearGradientAxis>, E: Var<ExtendMode>, O: Var<Point>, T: Var<Size>, TS: Var<Size>,

pub fn tile_spacing<TS2>( self, spacing: TS2, ) -> TiledLinearGradient<S, A, E, O, T, TS2::Var>
where TS2: IntoVar<Size>,

Set the space between tiles.

Relative values are resolved on the tile size, so setting this to 100.pct() will skip a tile.

Leftover values are resolved on the space taken by tiles that do not fully fit in the available space, so setting this to 1.lft() will cause the border tiles to always touch the full bounds and the middle filled with the maximum full tiles that fit or empty space.

pub fn tile_origin<O2>( self, origin: O2, ) -> TiledLinearGradient<S, A, E, O2::Var, T, TS>
where O2: IntoVar<Point>,

Sets the tile offset.

Relative values are resolved on the tile size, so setting this to 100.pct() will offset a full turn.